Create namespaceSelector.yaml
Description: The above definition file allows ingress traffic on db pods from api pods in all the namespaces. Assume we have multiple namespaces like prod, test, dev etc and we want to allow ingress traffic on ‘db’ pods only from ‘api’ pods in prod namespace.

Solution: For this we can make use of ‘namespaceSelector’ in the definition file. For namespace selector to work we need to label the namespace first